Minor code quality improvements
- group.cpp: use std::ptrdiff_t for iterator arithmetic in AddChild
- tests/CMakeLists.txt: replace deprecated LINK_FLAGS property with target_link_options
- test_interp.cpp: make InterpFuncs map test exhaustive — check all enum
  values via magic_enum::enum_count and verify every function pointer is
  non-null, rather than spot-checking 10 specific entries

+++ src/nodes/group.cpp
@@ -53,7 +53,7 @@ namespace moth_ui {
 
     void Group::AddChild(std::shared_ptr<Node> child, size_t index) {
         if (index != static_cast<size_t>(-1)) {
-            auto it = std::begin(m_children) + static_cast<int32_t>(index);
+            auto it = std::begin(m_children) + static_cast<std::ptrdiff_t>(index);
             m_children.insert(it, child);
         } else {
             m_children.push_back(child);
